Renamed wordpress -> wordpressify.
Depending on the gem wordpress-release broke the command 'wordpress'
by introducing a directory named 'wordpress' in the load path *before*
the executable 'wordpress'.

Renaming the executable was the simplest workaround I came up with,
though I'd love to find something simpler that would allow me to
resume using 'wordpress' as the executable name while still keeping
the internal structure of the gems nice.